requirements.txt
fastapi
uvicorn
langchain
langgraph
pandas
numpy
python-multipart
websockets
langchain-google-genai
python-dotenv
